+++ Source/SwiftLintFramework/Models/RuleParameter.swift
@@ -6,7 +6,7 @@
 //  Copyright (c) 2015 Realm. All rights reserved.
 //
 
-public struct RuleParameter<T> {
+public struct RuleParameter<T: Equatable> : Equatable {
     public let severity: ViolationSeverity
     public let value: T
 
@@ -14,4 +14,10 @@ public struct RuleParameter<T> {
         self.severity = severity
         self.value = value
     }
+}
+
+// MARK: - Equatable
+
+public func ==<T: Equatable>(lhs: RuleParameter<T>, rhs: RuleParameter<T>) -> Bool {
+    return lhs.value == rhs.value && lhs.severity == rhs.severity
 }