+++ Source/swiftlint/Extensions/Configuration+CommandLine.swift
@@ -80,7 +80,7 @@ extension Configuration {
                         let filename = (path as NSString).lastPathComponent
                         queuedPrintError("\(action) '\(filename)' (\(index + 1)/\(fileCount))")
                     }
-                    visitorBlock(Linter(file: file, configuration: self))
+                    visitorBlock(Linter(file: file, configuration: configForFile(file)))
                 }
                 return .Success(files)
         }