+++ CHANGELOG.md
@@ -18,6 +18,10 @@
   [JP Simard](https://github.com/jpsim)
   [#222](https://github.com/realm/SwiftLint/issues/222)
 
+* Add nested `.swiftlint.yml` configuration support.  
+  [Scott Hoyt](https://github.com/scottrhoyt)
+  [#299](https://github.com/realm/SwiftLint/issues/299)
+
 ##### Bug Fixes
 
 * Fix multibyte handling in many rules.